restructured and renamed the repo
