--- 2D_Classifier/train/requirements.txt
@@ -1,10 +1,0 @@
-csc-mlops
-torch
-torchmetrics
-torchvision
-pytorch_lightning
-ray
-timm
-pytest
-pytest-cov
-flake8
